Remove color_bar and color_bar_legend components
