+++ lib/phoenix/live_dashboard/components/shared_usage_card_component.ex
@@ -78,10 +78,36 @@ defmodule Phoenix.LiveDashboard.SharedUsageCardComponent do
         <div phx-hook="PhxColorBarHighlight" id="cpu-color-bars">
           <%= for usage <- @usages do %>
             <div class="flex-grow-1 mb-3">
-              <%= live_component @socket, ColorBarComponent, dom_id: "cpu-#{usage.dom_sub_id}", data: usage.data, title: usage.title, csp_nonces: @csp_nonces %>
+              <div class="progress color-bar-progress flex-grow-1 mb-3">
+                <span class="color-bar-progress-title"><%= usage.title %></span>
+                <%= for {{name, value, color, _desc}, index} <- Enum.with_index(usage.data) do %>
+                  <style nonce="<%= @csp_nonces.style %>">#<%= "cpu-#{usage.dom_sub_id}-progress-#{index}" %>{width:<%= value %>%}</style>
+                  <div
+                  title="<%= name %> - <%= format_percent(value) %>"
+                  class="progress-bar color-bar-progress-bar bg-gradient-<%= color %>"
+                  role="progressbar"
+                  aria-valuenow="<%= maybe_round(value) %>"
+                  aria-valuemin="0"
+                  aria-valuemax="100"
+                  data-name="<%= name %>"
+                  data-empty="<%= empty?(value) %>"
+                  id="<%= "cpu-#{usage.dom_sub_id}-progress-#{index}" %>">
+                  </div>
+                <% end %>
+              </div>
             </div>
           <% end %>
-          <%= live_component @socket, ColorBarLegendComponent, data: @total_data, formatter: @total_formatter %>
+          <div class="color-bar-legend">
+            <div class="row">
+            <%= for {name, value, color, hint} <- @total_data do %>
+              <div class="col-lg-6 d-flex align-items-center py-1 flex-grow-0 color-bar-legend-entry" data-name="<%= name %>">
+                <div class="color-bar-legend-color bg-<%= color %> mr-2"></div>
+                <span><%= name %> <%= hint && hint(do: hint) %></span>
+                <span class="flex-grow-1 text-right text-muted"><%= @total_formatter.(value) %></span>
+              </div>
+              <% end %>
+            </div>
+          </div>
           <div class="resource-usage-total text-center py-1 mt-3">
             <%= @total_legend %> <%= @total_usage %>
           </div>
@@ -90,4 +116,10 @@ defmodule Phoenix.LiveDashboard.SharedUsageCardComponent do
     </div>
     """
   end
+
+  defp maybe_round(num) when is_integer(num), do: num
+  defp maybe_round(num), do: Float.ceil(num, 1)
+
+  defp empty?(value) when is_number(value) and value > 0, do: false
+  defp empty?(_), do: true
 end
