+++ lib/phoenix/live_dashboard/pages/ets_page.ex
@@ -10,7 +10,7 @@ defmodule Phoenix.LiveDashboard.EtsPage do
   @impl true
   def render_page(_assigns) do
     table(
-      columns: columns(),
+      columns: table_columns(),
       id: @table_id,
       row_attrs: &row_attrs/1,
       row_fetcher: &fetch_ets/2,
@@ -25,7 +25,7 @@ defmodule Phoenix.LiveDashboard.EtsPage do
     SystemInfo.fetch_ets(node, search, sort_by, sort_dir, limit)
   end
 
-  defp columns() do
+  defp table_columns() do
     [
       %{
         field: :name,
